% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{as_quarto_typst}
\alias{as_quarto_typst}
\title{Raw blocks for Quarto}
\usage{
as_quarto_typst(x)
}
\arguments{
\item{x}{A character vector to interpret as raw typst.}
}
\value{
A raw version, wrapped in the expected format
for Quarto document.
}
\description{
Convert to raw Typst for Quarto.
Using this function is required to include the
output of \code{{r2typ}} in your document.
}
\note{
\code{{r2typ}} does not have any dependency, but this function tries
to reproduce the behavior \code{knitr::raw_block()} for Typst outputs.
}
\examples{
heading(level = 2, numbering = "1.1", "Hello world") |> as_quarto_typst()

typ_markup <- heading(level = 2, numbering = "1.1", "Hello world")
as_quarto_typst(typ_markup)
}
