% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\docType{data}
\name{pos_bracket_functions}
\alias{pos_bracket_functions}
\alias{no_bracket_functions}
\title{Functions with brackets.}
\format{
An object of class \code{character} of length 3.

An object of class \code{character} of length 31.
}
\usage{
pos_bracket_functions

no_bracket_functions
}
\description{
Function where positional arguments are in brackets.

Function where additional arguments is not content.
For example, image() usage must be image(height: 2in, "file.svg").
}
\keyword{internal}
