% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{format_as_typst}
\alias{format_as_typst}
\title{(internal) Format into Typst format}
\usage{
format_as_typst(x, named)
}
\arguments{
\item{x}{The object to convert.}

\item{named}{Whether \code{x} is named or not.}
}
\value{
Formatted to Typst
}
\description{
Format \code{x} to its Typst equivalent. This function
is used internally to convert R data types to their
equivalent in Typst. It works recursively.

This function is \strong{not meant for end-users}, use at
your own risk.
}
\examples{
format_as_typst(c(1, 2, 3))
format_as_typst(list(1, 2, 3))
format_as_typst(list(a=1, b=2, c=3))
format_as_typst(pt(20))
format_as_typst(TRUE)
format_as_typst(FALSE)
format_as_typst(NULL)
format_as_typst("auto")

}
