% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid.R
\name{is_valid_typst}
\alias{is_valid_typst}
\title{Is valid Typst?}
\usage{
is_valid_typst(x, error_on_failure = FALSE)
}
\arguments{
\item{x}{A character vector}

\item{error_on_failure}{Whether to raise an error if
the code is invalid. Default to \code{FALSE}.}
}
\value{
Whether or not pdf output file exist.
}
\description{
Check that a character vector is valid Typst markup
by compiling it. If no error, it assumes the code
is valid.
}
\note{
It requires to have the Typst compiler installed.
See https://typst.app/open-source/
}
\examples{
typst_code <- c("= Hello World", "This is a Typst document.")
is_valid_typst(typst_code)

typst_code <- c("= Hello World", "#This is a Typst document.")
is_valid_typst(typst_code)

typst_code <- c("= Hello World", "#This is a Typst document.")
is_valid_typst(typst_code, error_on_failure = TRUE)

}
