% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{parse_typst_args}
\alias{parse_typst_args}
\title{Parse \code{...} arguments}
\usage{
parse_typst_args(name, ...)
}
\arguments{
\item{name}{Name of the function}

\item{...}{Any kind of named or not arguments.}
}
\value{
A list with function name, named args and unnamed args.
}
\description{
This functions parses user arguments
from \code{...}. The goal is to get which arguments
are named or not, convert to their Typst format
and return a list with parsed info.
}
