% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{quarto}
\alias{quarto}
\alias{as_quarto_typst}
\alias{knit_print.typst_markup}
\title{Raw blocks for Quarto and Custom knit_print}
\usage{
as_quarto_typst(x)

\method{knit_print}{typst_markup}(x, ...)
}
\arguments{
\item{x}{A character vector to interpret as raw typst or
an object of class \code{typst_markup}}

\item{...}{Ignored}
}
\value{
For \code{as_quarto_typst()}, a raw version wrapped in the
expected Quarto format. For \code{knit_print.typst_markup()}, output
from \code{as_quarto_typst()}.
}
\description{
Convert to raw Typst for Quarto.
Using this function is required to include the
output of \code{{r2typ}} in your document.

Define output behavior when "knit printing" a \code{typst_markup}
object with \code{knitr} (mostly with Quarto).

Learn more in the \href{../articles/using-r2typ-in-quarto-documents.html}{Quarto usage vignette}.
}
\examples{
heading(level = 2, numbering = "1.1", "Hello world") |> as_quarto_typst()

typ_markup <- heading(level = 2, numbering = "1.1", "Hello world")
as_quarto_typst(typ_markup)

}
