% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_layout.R, R/functions_model.R,
%   R/functions_text.R, R/functions_visualize.R, R/show.R
\name{show_place}
\alias{show_place}
\alias{show_align}
\alias{show_h}
\alias{show_v}
\alias{show_block}
\alias{show_box}
\alias{show_colbreak}
\alias{show_stack}
\alias{show_grid}
\alias{show_hide}
\alias{show_pagebreak}
\alias{show_move}
\alias{show_pad}
\alias{show_page}
\alias{show_scale}
\alias{show_skew}
\alias{show_rotate}
\alias{show_bibliography}
\alias{show_list}
\alias{show_document}
\alias{show_emph}
\alias{show_figure}
\alias{show_footnote}
\alias{show_heading}
\alias{show_enum}
\alias{show_outline}
\alias{show_par}
\alias{show_parbreak}
\alias{show_quote}
\alias{show_strong}
\alias{show_table}
\alias{show_text}
\alias{show_linebreak}
\alias{show_overline}
\alias{show_smallcaps}
\alias{show_smartquote}
\alias{show_strike}
\alias{show_sub}
\alias{show_super}
\alias{show_underline}
\alias{show_upper}
\alias{show_raw}
\alias{show_highlight}
\alias{show_lower}
\alias{show_image}
\alias{show_circle}
\alias{show_ellipse}
\alias{show_line}
\alias{show_rect}
\alias{show_square}
\alias{typst-show}
\title{Typst show rules}
\usage{
show_place(set_rule)

show_align(set_rule)

show_h(set_rule)

show_v(set_rule)

show_block(set_rule)

show_box(set_rule)

show_colbreak(set_rule)

show_stack(set_rule)

show_grid(set_rule)

show_hide(set_rule)

show_pagebreak(set_rule)

show_move(set_rule)

show_pad(set_rule)

show_page(set_rule)

show_scale(set_rule)

show_skew(set_rule)

show_rotate(set_rule)

show_bibliography(set_rule)

show_list(set_rule)

show_document(set_rule)

show_emph(set_rule)

show_figure(set_rule)

show_footnote(set_rule)

show_heading(set_rule)

show_enum(set_rule)

show_outline(set_rule)

show_par(set_rule)

show_parbreak(set_rule)

show_quote(set_rule)

show_strong(set_rule)

show_table(set_rule)

show_text(set_rule)

show_linebreak(set_rule)

show_overline(set_rule)

show_smallcaps(set_rule)

show_smartquote(set_rule)

show_strike(set_rule)

show_sub(set_rule)

show_super(set_rule)

show_underline(set_rule)

show_upper(set_rule)

show_raw(set_rule)

show_highlight(set_rule)

show_lower(set_rule)

show_image(set_rule)

show_circle(set_rule)

show_ellipse(set_rule)

show_line(set_rule)

show_rect(set_rule)

show_square(set_rule)
}
\arguments{
\item{set_rule}{A set rule made with one of the set functions in \code{{r2typ}}.}
}
\value{
A character string containing the generated Typst code.
}
\description{
Generates Typst show rules for a given function.
}
