% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\docType{data}
\name{typst_auto}
\alias{typst_auto}
\alias{auto}
\title{Typst auto}
\format{
An object of class \code{typst_auto} (inherits from \code{typst_expression}) of length 1.
}
\usage{
auto
}
\value{
An object of class \code{typst_auto} and \code{typst_expression}.
}
\description{
The \code{auto} keyword in Typst.
}
\examples{
auto

}
\keyword{datasets}
