% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{typst_function}
\alias{typst_function}
\title{Typst function generator}
\usage{
typst_function(name, ...)
}
\arguments{
\item{name}{Function name like \code{heading()} or \code{text()}}

\item{...}{Additional arguments pasted to typst functions.}
}
\description{
Base function used to generate
higher level functions.
}
\keyword{internal}
