% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{typst_colors}
\alias{typst_colors}
\alias{rgb_}
\alias{silver}
\alias{grey}
\alias{black}
\alias{white}
\alias{navy}
\alias{blue}
\alias{aqua}
\alias{teal}
\alias{eastern}
\alias{purple}
\alias{fuchsia}
\alias{maroon}
\alias{red}
\alias{orange}
\alias{yellow}
\alias{olive}
\alias{green}
\alias{lime}
\title{Typst Color}
\format{
An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_color} (inherits from \code{typst_expression}) of length 1.
}
\usage{
rgb_(hex)

silver

grey

black

white

navy

blue

aqua

teal

eastern

purple

fuchsia

maroon

red

orange

yellow

olive

green

lime
}
\arguments{
\item{hex}{An hexadecimal color.}
}
\value{
An object of class \code{typst_color} and \code{typst_expression}.
}
\description{
These functions generate special objects representing colors in Typst code.
They can be used as arguments in functions like \code{text_()} or \code{heading()} to
produce properly formatted Typst code.
}
\examples{
rgb_("#ffffff")
text_(font = "Roboto", fill = rgb_("#ffffff"), "cookies are healthy!!")
red
blue
black

}
\keyword{datasets}
