% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{typst_compile}
\alias{typst_compile}
\title{Compile a Typst file or character vector}
\usage{
typst_compile(file_or_chr, output = NULL)
}
\arguments{
\item{file_or_chr}{Path to a file (must end with ".typ") or a character vector.}

\item{output}{Optional path for the output file (default NULL).}
}
\description{
Call the Typst CLI compiler and compile a file (.typ)
or a character vector.
}
