% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{typst_compile}
\alias{typst_compile}
\title{Compile a \code{.typ} file to a \code{.pdf} file and return the output path}
\usage{
typst_compile(file, ...)
}
\arguments{
\item{file}{Path to an existing \code{.typ} file.}

\item{...}{Additional arguments passed to \code{\link[tynding:typst_compile]{tynding::typst_compile()}}}
}
\value{
The path to the compiled PDF file.
}
\description{
Compile a \code{.typ} file to a \code{.pdf} file and return the output path.
}
