% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\docType{data}
\name{typst_direction}
\alias{typst_direction}
\alias{ltr}
\alias{rtl}
\alias{ttb}
\alias{btt}
\title{Create Typst Direction}
\format{
An object of class \code{typst_direction} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_direction} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_direction} (inherits from \code{typst_expression}) of length 1.

An object of class \code{typst_direction} (inherits from \code{typst_expression}) of length 1.
}
\usage{
ltr

rtl

ttb

btt
}
\value{
An object of class \code{typst_direction} and \code{typst_expression}.
}
\description{
These functions generate special objects representing directions in Typst.
They can be used as arguments in functions like \code{stack()}.
}
\examples{
ltr
rtl
ttb
btt

}
\keyword{datasets}
