% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{typst_function}
\alias{typst_function}
\title{Create a Typst function}
\usage{
typst_function(name, ...)
}
\arguments{
\item{name}{Function name}

\item{...}{Any kind of named or not arguments.}
}
\value{
A character vector with well formatted Typst.
}
\description{
Used to create all Typst functions in \code{r2typ}.
Arguments are parsed and then translated to Typst.
}
\keyword{internal}
