% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{typst_show}
\alias{typst_show}
\title{Typst show rules}
\usage{
typst_show(name, set_rule)
}
\arguments{
\item{name}{Name of the function for the show rule.}

\item{set_rule}{A set rule made with one of the set functions in \code{{r2typ}}.}
}
\value{
A character string containing the generated Typst code.
}
\description{
Generates Typst show rules for a given function.
}
\keyword{internal}
