% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{typst_units}
\alias{typst_units}
\alias{pt}
\alias{em}
\alias{percent}
\title{Create Typst Unit Values}
\usage{
pt(x)

em(x)

percent(x)
}
\arguments{
\item{x}{Numeric value of the unit.}
}
\value{
An object of class \code{typst_unit} with an associated unit attribute.
When passed to a Typst-generating function, it is converted to the
appropriate unit string (e.g., \verb{12pt}, \verb{1.5em}, \verb{80\%}).
}
\description{
These functions generate special objects representing units in Typst code.
They can be used as arguments in functions like \code{text()} or \code{heading()} to
produce properly formatted Typst code.
}
\examples{
pt(12)       # Represents 12pt
em(1.5)      # Represents 1.5em
percent(80)  # Represents 80\%

text(size = pt(12), "Hello world")

}
