% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{typst_write}
\alias{typst_write}
\title{Write a Typst file from a character vector}
\usage{
typst_write(x, ...)
}
\arguments{
\item{x}{A character vector representing Typst code.}

\item{...}{Additional arguments passed to \code{\link[tynding:typst_write]{tynding::typst_write()}}}
}
\value{
The path to the written .typ file, invisibly.
}
\description{
Create a Typst file (.typ) from a character vector.
}
\examples{
\dontrun{
code <- c("= Hello World", "This is a Typst document.")
typst_write(code, output = "hello.typ")
}

}
