% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{as_raw_typ}
\alias{as_raw_typ}
\title{Raw blocks for Rmarkdown/Quarto}
\usage{
as_raw_typ(x)
}
\arguments{
\item{x}{A character vector to interpret as raw typst.}
}
\value{
A raw version, wrapped in the expected format
for Rmarkdown and Quarto document.
}
\description{
Convert to raw Typst for Rmarkdown and Quarto.
Using this function is required to include the
output of \code{{rtyp}} in your document.

This is a simple wrapper around \code{\link[knitr:raw_block]{knitr::raw_block()}}.
}
\examples{
heading(level = 2, numbering = "1.1", "Hello world") |> as_raw_typ()

typ_markup <- heading(level = 2, numbering = "1.1", "Hello world")
as_raw_typ(typ_markup)
}
