% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\docType{data}
\name{expression_like}
\alias{expression_like}
\alias{pos_bracket_functions}
\alias{no_bracket_functions}
\title{Expression like classes/}
\format{
An object of class \code{character} of length 4.

An object of class \code{character} of length 3.

An object of class \code{character} of length 31.
}
\usage{
expression_like

pos_bracket_functions

no_bracket_functions
}
\description{
Class names of expressions like.

Function where positional arguments are in brackets.

Function where additional arguments is not content.
For example, image() usage must be image(height: 2in, "file.svg").
}
\keyword{internal}
