% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/let.R
\name{let}
\alias{let}
\title{Typst let keyword}
\usage{
let(name, value)
}
\arguments{
\item{name}{Name of the variables.}

\item{value}{Value to assign to \code{name}.}
}
\value{
A character string
}
\description{
\code{let} is the keyword used to define variables in Typst.
It allows to create reusable components.

Note that it will call \code{assign(name, value, envir = parent.frame())}
so that \code{name} becomes an actual variable in your R environment.
}
\examples{
let("yellow", rgb("#FFC300"))

let("mycirc", circle(width=cm(3), fill=yellow, square(height=cm(1))))

}
