% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\docType{data}
\name{typst_alignments}
\alias{typst_alignments}
\alias{+.typst_alignment}
\alias{horizon}
\alias{top}
\alias{bottom}
\alias{center}
\alias{left}
\alias{right}
\title{Create Typst Alignment}
\format{
An object of class \code{typst_alignment} of length 1.

An object of class \code{typst_alignment} of length 1.

An object of class \code{typst_alignment} of length 1.

An object of class \code{typst_alignment} of length 1.

An object of class \code{typst_alignment} of length 1.

An object of class \code{typst_alignment} of length 1.
}
\usage{
\method{+}{typst_alignment}(e1, e2)

horizon

top

bottom

center

left

right
}
\arguments{
\item{e1}{A typst alignment}

\item{e2}{A typst alignment}
}
\value{
An object of class \code{typst_alignment}.
}
\description{
These functions generate special objects representing alignment in Typst code.
They can be used as arguments in functions like \code{text()} or \code{heading()} to
produce properly formatted Typst code.
}
\examples{
left
center + horizon
top + right

}
\keyword{datasets}
