% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{typst_compile}
\alias{typst_compile}
\title{Compile a \code{.typ} file to a sibling \code{.pdf} file and return the output path.}
\usage{
typst_compile(file)
}
\description{
Compile a \code{.typ} file to a sibling \code{.pdf} file and return the output path.
}
