% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_foundations.R,
%   R/functions_layout.R, R/functions_model.R, R/functions_text.R,
%   R/functions_visualize.R
\name{typst_functions}
\alias{typst_functions}
\alias{assert}
\alias{bytes}
\alias{datetime}
\alias{duration}
\alias{eval_}
\alias{panic}
\alias{regex}
\alias{repr}
\alias{str_}
\alias{symbol}
\alias{decimal}
\alias{type}
\alias{version}
\alias{place}
\alias{align}
\alias{h}
\alias{v}
\alias{block}
\alias{box}
\alias{colbreak}
\alias{stack}
\alias{grid}
\alias{hide}
\alias{pagebreak}
\alias{move}
\alias{pad}
\alias{page}
\alias{bibliography}
\alias{cite}
\alias{list_}
\alias{document}
\alias{emph}
\alias{figure}
\alias{footnote}
\alias{heading}
\alias{enum}
\alias{outline}
\alias{par}
\alias{parbreak}
\alias{quote_}
\alias{strong}
\alias{table_}
\alias{text}
\alias{linebreak}
\alias{highlight}
\alias{lower}
\alias{overline}
\alias{smallcaps}
\alias{smartquote}
\alias{strike}
\alias{sub_}
\alias{super}
\alias{underline}
\alias{upper}
\alias{raw_}
\alias{image}
\alias{circle}
\alias{ellipse}
\alias{line}
\alias{rect}
\alias{square}
\title{Typst function}
\usage{
assert(...)

bytes(...)

datetime(...)

duration(...)

eval_(...)

panic(...)

regex(...)

repr(...)

str_(...)

symbol(...)

decimal(...)

type(...)

version(...)

place(...)

align(...)

h(...)

v(...)

block(...)

box(...)

colbreak(...)

stack(...)

grid(...)

hide(...)

pagebreak(...)

move(...)

pad(...)

page(...)

bibliography(...)

cite(...)

list_(...)

document(...)

emph(...)

figure(...)

footnote(...)

heading(...)

enum(...)

outline(...)

par(...)

parbreak(...)

quote_(...)

strong(...)

table_(...)

text(...)

linebreak(...)

highlight(...)

lower(...)

overline(...)

smallcaps(...)

smartquote(...)

strike(...)

sub_(...)

super(...)

underline(...)

upper(...)

raw_(...)

image(...)

circle(...)

ellipse(...)

line(...)

rect(...)

square(...)
}
\arguments{
\item{...}{Named or unnamed arguments passed to the Typst function.
Named arguments become Typst parameters (e.g. \code{level = 2} → \code{level: 2}),
while unnamed arguments form the body inside \verb{[...]}.}
}
\value{
A character string containing the generated Typst code.
}
\description{
Generates Typst code with element. Named arguments are included
as Typst keyword arguments (e.g. \code{level}, \code{numbering}), and unnamed arguments
are used as the heading content.
But please note that this behavior might be different in some functions
to respect true Typst syntax (such as \verb{#table()} for instance).
}
\note{
Does not work yet because \code{label} type is not implemented yet.
See https://github.com/y-sunflower/rtyp/issues/10
}
\examples{
heading(level = 2, numbering = "1.1", "Hello world")
text(font = "Libertinus Serif", "Hello world")
text("No arguments")

assert(1 < 2, message = "math broke")
bytes("more cookies...")
datetime(year = 2012, month = 8, day = 3)
duration(days = 3, hours = 12)
eval_("*Markup!*", mode = "markup")
panic("this is wrong")
regex("\\\\d+")
repr("hello")
str_("hello")
symbol("10")
decimal("10")
type("10")
version(1, 2, 3, 4)
version(c(1, 2, 3, 4))
place(dx = pt(50), "pasta")
align(center + horizon, "hey")
h(fr(3))
v(percent(3), weak = TRUE)
block(breakable = FALSE, width = percent(50), height = "auto", fill = aqua, "cookies")
box(clip = FALSE, width = percent(50), height = "auto", fill = red, "ice cream")
colbreak(weak = TRUE)
stack(dir = ttb, "hey", "you", "!")
grid(columns = 2, rows = 2, gutter = pt(3), "hey", "cookies", "ice", "cream")
hide("this is hidden")
pagebreak("this is hidden")
move(dx = pt(3), dy = pt(5), "pasta")
pad(x = percent(0) + pt(5), y = percent(10) + pt(0), image("file.svg"))
page(
 flipped = TRUE,
 columns = 2,
 fill = red,
 place(
   top + left,
   dx = pt(-5),
   rect(fill = blue, radius = pt(2), "yooooo")
 )
)
highlight(extent = pt(3), "hey")
lower("hey")
raw_("print('hello world')", lang = "python", block = TRUE)
}
