% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_model.R, R/set.R
\name{set_bibliography}
\alias{set_bibliography}
\alias{set_document}
\alias{set_emph}
\alias{set_figure}
\alias{set_footnote}
\alias{set_heading}
\alias{set_enum}
\alias{set_outline}
\alias{set_par}
\alias{set_parbreak}
\alias{set_quote}
\alias{set_strong}
\alias{set_table}
\alias{typst_set}
\title{Typst set rules}
\usage{
set_bibliography(...)

set_document(...)

set_emph(...)

set_figure(...)

set_footnote(...)

set_heading(...)

set_enum(...)

set_outline(...)

set_par(...)

set_parbreak(...)

set_quote(...)

set_strong(...)

set_table(...)
}
\arguments{
\item{...}{Named or unnamed arguments passed to the Typst function.
Named arguments become Typst parameters (e.g. \code{level = 2} → \code{level: 2}),
while unnamed arguments form the body inside \verb{[...]}.}
}
\value{
A character string containing the generated Typst code.
}
\description{
Generates Typst set rules for a given function.
}
\examples{
set_bibliography(full = TRUE)
set_document(author = "Me, myself & I")
set_emph()
set_figure(alt = "This is an alternative text")
set_footnote(numbering = "*")
set_heading(level = 3)
set_enum(tight = FALSE)
set_outline(depth = 1)
set_par(
 `first-line-indent` = em(1),
 spacing = em(0.65),
 justify = TRUE
)
set_quote(block = TRUE)
set_strong(delta = 400)
set_table(align = center, inset = pt(10))
}
