% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_layout.R, R/functions_model.R,
%   R/set.R
\name{set_place}
\alias{set_place}
\alias{set_align}
\alias{set_h}
\alias{set_v}
\alias{set_block}
\alias{set_box}
\alias{set_colbreak}
\alias{set_stack}
\alias{set_grid}
\alias{set_hide}
\alias{set_pagebreak}
\alias{set_move}
\alias{set_pad}
\alias{set_page}
\alias{set_scale}
\alias{set_skew}
\alias{set_bibliography}
\alias{set_document}
\alias{set_emph}
\alias{set_figure}
\alias{set_footnote}
\alias{set_heading}
\alias{set_enum}
\alias{set_outline}
\alias{set_par}
\alias{set_parbreak}
\alias{set_quote}
\alias{set_strong}
\alias{set_table}
\alias{typst_set}
\title{Typst set rules}
\usage{
set_place(...)

set_align(...)

set_h(...)

set_v(...)

set_block(...)

set_box(...)

set_colbreak(...)

set_stack(...)

set_grid(...)

set_hide(...)

set_pagebreak(...)

set_move(...)

set_pad(...)

set_page(...)

set_scale(...)

set_skew(...)

set_bibliography(...)

set_document(...)

set_emph(...)

set_figure(...)

set_footnote(...)

set_heading(...)

set_enum(...)

set_outline(...)

set_par(...)

set_parbreak(...)

set_quote(...)

set_strong(...)

set_table(...)
}
\arguments{
\item{...}{Named or unnamed arguments passed to the Typst function.
Named arguments become Typst parameters (e.g. \code{level = 2} → \code{level: 2}),
while unnamed arguments form the body inside \verb{[...]}.}
}
\value{
A character string containing the generated Typst code.
}
\description{
Generates Typst set rules for a given function.
}
\examples{
set_place(dx = pt(50))
set_align(center + horizon)
set_h(weak = TRUE)
set_v(weak = TRUE)
set_block(breakable = FALSE, width = percent(50), height = "auto", fill = aqua)
set_box(clip = FALSE, width = percent(50), height = "auto", fill = red)
set_colbreak(weak = TRUE)
set_stack(dir = ttb)
set_grid(columns = 2, rows = 2, gutter = pt(3))
set_pagebreak(weak = TRUE)
set_move(dx = pt(3), dy = pt(5))
set_pad(x = percent(0) + pt(5), y = percent(10) + pt(0))
set_page(
 flipped = TRUE,
 columns = 2,
 fill = red
)
set_scale(x = percent(-100))
set_skew(ax = deg(-12))
set_bibliography(full = TRUE)
set_document(author = "Me, myself & I")
set_emph()
set_figure(alt = "This is an alternative text")
set_footnote(numbering = "*")
set_heading(level = 3)
set_enum(tight = FALSE)
set_outline(depth = 1)
set_par(
 `first-line-indent` = em(1),
 spacing = em(0.65),
 justify = TRUE
)
set_quote(block = TRUE)
set_strong(delta = 400)
set_table(align = center, inset = pt(10))
}
