% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{typst_show}
\alias{typst_show}
\alias{show_}
\title{Typst show rules}
\usage{
show_(name, set_rule)
}
\arguments{
\item{name}{Name of the function for the show rule.}

\item{set_rule}{A set rule made with one of the set functions
in \code{{r2typ}}.}
}
\value{
A character string containing the generated Typst code.
}
\description{
Generates Typst show rules for a given function.
}
\examples{
show_("heading", set_text(red))

}
